status - move tests
