+++ src/status_test.rs
@@ -2,15 +2,15 @@ use super::*;
 use crate::sentry_api::SentryApi;
 use chrono::{Duration, Utc};
 use primitives::{
-    sentry::{LastApproved, LastApprovedResponse, NewStateValidatorMessage},
-    validator::{Heartbeat, MessageTypes, NewState},
+    sentry::{
+        ApproveStateValidatorMessage, LastApproved, LastApprovedResponse, NewStateValidatorMessage,
+    },
+    util::tests::prep_db::{DUMMY_CHANNEL, DUMMY_VALIDATOR_FOLLOWER, DUMMY_VALIDATOR_LEADER, IDS},
+    validator::{ApproveState, Heartbeat, MessageTypes, NewState},
     BalancesMap, Channel,
 };
 
 use httptest::{mappers::*, responders::*, Expectation, Server, ServerPool};
-use primitives::util::tests::prep_db::{
-    DUMMY_CHANNEL, DUMMY_VALIDATOR_FOLLOWER, DUMMY_VALIDATOR_LEADER,
-};
 
 static SERVER_POOL: ServerPool = ServerPool::new(4);
 
@@ -35,6 +35,30 @@ fn get_heartbeat_msg(recency: Duration, from: ValidatorId) -> HeartbeatValidator
             signature: String::new(),
             state_root: String::new(),
             timestamp: Utc::now() - recency,
+        }),
+    }
+}
+
+fn get_approve_state_msg() -> ApproveStateValidatorMessage {
+    ApproveStateValidatorMessage {
+        from: DUMMY_VALIDATOR_LEADER.id,
+        received: Utc::now(),
+        msg: MessageTypes::ApproveState(ApproveState {
+            state_root: String::from("0x0"),
+            signature: String::from("0x0"),
+            is_healthy: true,
+        }),
+    }
+}
+
+fn get_new_state_msg() -> NewStateValidatorMessage {
+    NewStateValidatorMessage {
+        from: DUMMY_VALIDATOR_LEADER.id,
+        received: Utc::now(),
+        msg: MessageTypes::NewState(NewState {
+            signature: String::from("0x0"),
+            state_root: String::from("0x0"),
+            balances: Default::default(),
         }),
     }
 }
@@ -211,7 +235,6 @@ mod is_offline {
                 last_approved: None,
                 heartbeats: follower,
             },
-            // we don't check the Recency, so we don't care about it's value
             recency: *RECENCY,
         }
     }
@@ -286,5 +309,351 @@ mod is_offline {
             false,
             "Not offline: If it has recent heartbeats in both Leader & Follower!"
         );
+    }
+}
+
+mod is_date_recent {
+    use super::*;
+
+    #[test]
+    fn now_date_is_recent() {
+        let now = Utc::now();
+        let recency = Duration::minutes(4);
+        assert!(
+            is_date_recent(recency, &now),
+            "The present moment is a recent date!"
+        )
+    }
+
+    #[test]
+    fn slightly_past_is_recent() {
+        let recency = Duration::minutes(4);
+        let on_the_edge = Utc::now() - Duration::minutes(3);
+        assert!(
+            is_date_recent(recency, &on_the_edge),
+            "When date is just as old as the recency limit, it still counts as recent"
+        )
+    }
+
+    #[test]
+    fn old_date_is_not_recent() {
+        let recency = Duration::minutes(4);
+        let past = Utc::now() - Duration::minutes(10);
+        assert_eq!(
+            is_date_recent(recency, &past),
+            false,
+            "Date older than the recency limit is not recent"
+        )
+    }
+}
+
+mod is_initializing {
+    use super::*;
+
+    #[test]
+    fn two_empty_message_arrays() {
+        let messages = Messages {
+            leader: LastApprovedResponse {
+                last_approved: None,
+                heartbeats: Some(vec![]),
+            },
+            follower: LastApprovedResponse {
+                last_approved: None,
+                heartbeats: Some(vec![]),
+            },
+            recency: Duration::minutes(4),
+        };
+
+        assert_eq!(
+            is_initializing(&messages),
+            true,
+            "Both leader heartbeats + newstate and follower heatbeats + approvestate pairs are empty arrays"
+        )
+    }
+
+    #[test]
+    fn leader_has_no_messages() {
+        let approve_state = get_approve_state_msg();
+        let heartbeat = get_heartbeat_msg(Duration::minutes(0), IDS["leader"]);
+        let messages = Messages {
+            leader: LastApprovedResponse {
+                last_approved: None,
+                heartbeats: Some(vec![]),
+            },
+            follower: LastApprovedResponse {
+                last_approved: Some(LastApproved {
+                    new_state: None,
+                    approve_state: Some(approve_state),
+                }),
+                heartbeats: Some(vec![heartbeat]),
+            },
+            recency: Duration::minutes(4),
+        };
+
+        assert_eq!(
+            is_initializing(&messages),
+            true,
+            "Leader has no new messages but the follower has heartbeats and approvestate"
+        )
+    }
+
+    #[test]
+    fn follower_has_no_messages() {
+        let heartbeat = get_heartbeat_msg(Duration::minutes(0), IDS["leader"]);
+        let new_state = get_new_state_msg();
+
+        let messages = Messages {
+            leader: LastApprovedResponse {
+                last_approved: Some(LastApproved {
+                    new_state: Some(new_state),
+                    approve_state: None,
+                }),
+                heartbeats: Some(vec![heartbeat]),
+            },
+            follower: LastApprovedResponse {
+                last_approved: None,
+                heartbeats: Some(vec![]),
+            },
+            recency: Duration::minutes(4),
+        };
+
+        assert_eq!(
+            is_initializing(&messages),
+            true,
+            "Follower has no new messages but leader has heartbeats and newstate"
+        )
+    }
+
+    #[test]
+    fn both_arrays_have_messages() {
+        let leader_heartbeats = vec![get_heartbeat_msg(Duration::minutes(0), IDS["leader"])];
+        let follower_heartbeats = vec![get_heartbeat_msg(Duration::minutes(0), IDS["leader"])];
+        let new_state = get_new_state_msg();
+        let approve_state = get_approve_state_msg();
+
+        let messages = Messages {
+            leader: LastApprovedResponse {
+                last_approved: Some(LastApproved {
+                    new_state: Some(new_state),
+                    approve_state: None,
+                }),
+                heartbeats: Some(leader_heartbeats),
+            },
+            follower: LastApprovedResponse {
+                last_approved: Some(LastApproved {
+                    new_state: None,
+                    approve_state: Some(approve_state),
+                }),
+                heartbeats: Some(follower_heartbeats),
+            },
+            recency: Duration::minutes(4),
+        };
+
+        assert_eq!(
+            is_initializing(&messages),
+            false,
+            "Both arrays have messages"
+        )
+    }
+}
+
+mod is_disconnected {
+    use super::*;
+
+    #[test]
+    fn no_recent_hbs_on_both_sides() {
+        let channel = DUMMY_CHANNEL.clone();
+        let leader_heartbeats = vec![get_heartbeat_msg(
+            Duration::minutes(10),
+            channel.spec.validators.leader().id,
+        )];
+        let follower_heartbeats = vec![get_heartbeat_msg(
+            Duration::minutes(10),
+            channel.spec.validators.leader().id,
+        )];
+
+        let messages = Messages {
+            leader: LastApprovedResponse {
+                last_approved: Some(LastApproved {
+                    new_state: None,
+                    approve_state: None,
+                }),
+                heartbeats: Some(leader_heartbeats),
+            },
+            follower: LastApprovedResponse {
+                last_approved: Some(LastApproved {
+                    new_state: None,
+                    approve_state: None,
+                }),
+                heartbeats: Some(follower_heartbeats),
+            },
+            recency: Duration::minutes(4),
+        };
+
+        assert_eq!(
+            is_disconnected(&channel, &messages),
+            true,
+            "Both leader and follower heartbeats have no recent messages"
+        )
+    }
+
+    #[test]
+    fn no_recent_follower_hbs() {
+        let channel = DUMMY_CHANNEL.clone();
+        let leader_heartbeats = vec![
+            get_heartbeat_msg(Duration::minutes(10), channel.spec.validators.leader().id),
+            get_heartbeat_msg(Duration::minutes(10), channel.spec.validators.follower().id),
+        ];
+        let follower_heartbeats = vec![
+            get_heartbeat_msg(Duration::minutes(10), channel.spec.validators.leader().id),
+            get_heartbeat_msg(Duration::minutes(10), channel.spec.validators.follower().id),
+        ];
+
+        let new_state = get_new_state_msg();
+        let approve_state = get_approve_state_msg();
+
+        let messages = Messages {
+            leader: LastApprovedResponse {
+                last_approved: Some(LastApproved {
+                    new_state: Some(new_state),
+                    approve_state: None,
+                }),
+                heartbeats: Some(leader_heartbeats),
+            },
+            follower: LastApprovedResponse {
+                last_approved: Some(LastApproved {
+                    new_state: None,
+                    approve_state: Some(approve_state),
+                }),
+                heartbeats: Some(follower_heartbeats),
+            },
+            recency: Duration::minutes(4),
+        };
+
+        assert_eq!(
+            is_disconnected(&channel, &messages),
+            true,
+            "No recent heartbeats on both validators"
+        )
+    }
+
+    #[test]
+    fn no_hb_in_leader_where_from_points_to_follower() {
+        let channel = DUMMY_CHANNEL.clone();
+        let leader_heartbeats = vec![get_heartbeat_msg(
+            Duration::minutes(0),
+            channel.spec.validators.leader().id,
+        )];
+        let follower_heartbeats = vec![
+            get_heartbeat_msg(Duration::minutes(0), channel.spec.validators.leader().id),
+            get_heartbeat_msg(Duration::minutes(0), channel.spec.validators.follower().id),
+        ];
+
+        let new_state = get_new_state_msg();
+        let approve_state = get_approve_state_msg();
+
+        let messages = Messages {
+            leader: LastApprovedResponse {
+                last_approved: Some(LastApproved {
+                    new_state: Some(new_state),
+                    approve_state: None,
+                }),
+                heartbeats: Some(leader_heartbeats),
+            },
+            follower: LastApprovedResponse {
+                last_approved: Some(LastApproved {
+                    new_state: None,
+                    approve_state: Some(approve_state),
+                }),
+                heartbeats: Some(follower_heartbeats),
+            },
+            recency: Duration::minutes(4),
+        };
+
+        assert_eq!(
+            is_disconnected(&channel, &messages),
+            true,
+            "Leader validator heartbeats have no recent messages that came from the follower"
+        )
+    }
+
+    #[test]
+    fn no_hb_in_follower_where_from_points_to_leader() {
+        let channel = DUMMY_CHANNEL.clone();
+        let leader_heartbeats = vec![
+            get_heartbeat_msg(Duration::minutes(0), channel.spec.validators.leader().id),
+            get_heartbeat_msg(Duration::minutes(0), channel.spec.validators.leader().id),
+        ];
+        let follower_heartbeats = vec![get_heartbeat_msg(
+            Duration::minutes(0),
+            channel.spec.validators.follower().id,
+        )];
+
+        let new_state = get_new_state_msg();
+        let approve_state = get_approve_state_msg();
+
+        let messages = Messages {
+            leader: LastApprovedResponse {
+                last_approved: Some(LastApproved {
+                    new_state: Some(new_state),
+                    approve_state: None,
+                }),
+                heartbeats: Some(leader_heartbeats),
+            },
+            follower: LastApprovedResponse {
+                last_approved: Some(LastApproved {
+                    new_state: None,
+                    approve_state: Some(approve_state),
+                }),
+                heartbeats: Some(follower_heartbeats),
+            },
+            recency: Duration::minutes(4),
+        };
+
+        assert_eq!(
+            is_disconnected(&channel, &messages),
+            true,
+            "Follower validator heartbeats have no recent messages that came from the leader"
+        )
+    }
+
+    #[test]
+    fn recent_hbs_coming_from_both_validators() {
+        let channel = DUMMY_CHANNEL.clone();
+        let leader_heartbeats = vec![
+            get_heartbeat_msg(Duration::minutes(0), channel.spec.validators.leader().id),
+            get_heartbeat_msg(Duration::minutes(0), channel.spec.validators.follower().id),
+        ];
+        let follower_heartbeats = vec![
+            get_heartbeat_msg(Duration::minutes(0), channel.spec.validators.leader().id),
+            get_heartbeat_msg(Duration::minutes(0), channel.spec.validators.follower().id),
+        ];
+
+        let new_state = get_new_state_msg();
+        let approve_state = get_approve_state_msg();
+
+        let messages = Messages {
+            leader: LastApprovedResponse {
+                last_approved: Some(LastApproved {
+                    new_state: Some(new_state),
+                    approve_state: None,
+                }),
+                heartbeats: Some(leader_heartbeats),
+            },
+            follower: LastApprovedResponse {
+                last_approved: Some(LastApproved {
+                    new_state: None,
+                    approve_state: Some(approve_state),
+                }),
+                heartbeats: Some(follower_heartbeats),
+            },
+            recency: Duration::minutes(4),
+        };
+
+        assert_eq!(
+            is_disconnected(&channel, &messages),
+            false,
+            "Leader hb has recent messages that came from the follower, and the follower has recent messages that came from the leader"
+        )
     }
 }
